requirements.txt
anyio==4.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
geographiclib==2.0
geopy==2.4.1
google-api-core==2.25.1
google-api-python-client==2.144.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
oauthlib==3.3.1
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-telegram-bot==22.3
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
uritemplate==4.2.0
urllib3==2.5.0
Flask==3.0.3
Werkzeug==3.1.3
python-dotenv==1.0.0